+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add id for Pane component. Refs UICHKOUT-768.
 * Add pull request template. Refs UICHKOUT-771.
 * Compile Translation Files into AST Format. Refs UICHKOUT-708.
+* Remove unnecessary request to `loan-policy-storage/loan-policies` end-point within checkout procedure. Refs UICHKOUT-767.
 
 ## [8.0.0](https://github.com/folio-org/ui-checkout/tree/v8.0.0) (2022-02-24)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v7.1.0...v8.0.0)